added FilterProvider
