+++ frontend/src/pages/ListProducts/index.tsx
@@ -11,30 +11,14 @@ import Stack from '@mui/material/Stack'
 import ProductCard from '../../components/ProductCard'
 import TableHeader from './TableHeader'
 import ProductCardList from './ProductCardList'
+import ListProductsInternal from './ListProducts'
+import { FilterProvider } from '../../providers/useFetchProducts'
 
 const ListProducts = () => {
-    const theme = useTheme()
-    const { data, loading, error } = useFetchPosts()
-
     return (
-        <Container sx={{ py: 5 }}>
-            <Stack spacing={3} sx={{ pt: 10 }}>
-                <TableHeader theme={theme} />
-
-                <Box
-                    display="flex"
-                    flexWrap="wrap"
-                    gap={3}
-                    justifyContent="center"
-                >
-                    {loading ? (
-                        <p style={{ color: '#ffffff' }}>Loading...</p>
-                    ) : (
-                        <ProductCardList data={data} />
-                    )}
-                </Box>
-            </Stack>
-        </Container>
+        <FilterProvider>
+            <ListProductsInternal />
+        </FilterProvider>
     )
 }
